Shopping cart

+++ shoppingcart/src/main/java/com/tdd/shoppingcart/offer/NoOffer.java
@@ -3,7 +3,7 @@ package com.tdd.shoppingcart.offer;
 public class NoOffer implements Offer {
 
     @Override
-    public int getPayableQuantity(int totalQuantity) {
+    public double getPayableQuantity(int totalQuantity) {
         return totalQuantity;
     }
 }
