+++ shoppingcart/src/main/java/com/tdd/shoppingcart/offer/Offer.java
@@ -1,5 +1,5 @@
 package com.tdd.shoppingcart.offer;
 
 public interface Offer {
-    int getPayableQuantity(int totalQuantity);
+    double getPayableQuantity(int totalQuantity);
 }
